+++ app/scripts/fhir/sdc-export-common.js
@@ -80,6 +80,191 @@ function addCommonSDCExportFns(ns) {
 
 
   /**
+   * Process an item of the form
+   * @param item an item in LForms form object
+   * @param source a LForms form object
+   * @param noExtensions a flag that a standard FHIR Questionnaire is to be created without any extensions.
+   *        The default is false.
+   * @returns {{}}
+   * @private
+   */
+  self._processItem = function(item, source, noExtensions) {
+    var targetItem = {};
+
+    // type
+    targetItem.type = this._getFhirDataType(item);
+
+    // id (empty for new record)
+
+    // code
+    targetItem.code = item.codeList;
+
+    // extension
+    targetItem.extension = [];
+
+    // required
+    targetItem.required = item._answerRequired;
+
+    // http://hl7.org/fhir/StructureDefinition/questionnaire-minOccurs
+    if (targetItem.required) {
+      targetItem.extension.push({
+        "url" : "http://hl7.org/fhir/StructureDefinition/questionnaire-minOccurs",
+        "valueInteger" : parseInt(item.questionCardinality.min)
+      });
+    }
+
+    // question repeats
+    // http://hl7.org/fhir/StructureDefinition/questionnaire-maxOccurs
+    this._processQuestionCardinality(targetItem, item);
+
+    // answer repeats
+    // http://hl7.org/fhir/StructureDefinition/questionnaire-maxOccurs
+    if (item.answerCardinality) {
+      if (item.answerCardinality.max === "*") {
+        targetItem.extension.push({
+          "url": "http://hl7.org/fhir/StructureDefinition/questionnaire-answerRepeats",
+          "valueBoolean": true
+        });
+      }
+    }
+
+    // Copied FHIRPath-related (pre-pop & extraction) extensions
+    this._processFHIRPathExtensions(targetItem, item);
+
+    // http://hl7.org/fhir/StructureDefinition/questionnaire-itemControl
+    this._handleItemControl(targetItem, item);
+
+    // check restrictions
+    this._handleRestrictions(targetItem, item);
+
+    // http://hl7.org/fhir/StructureDefinition/entryFormat
+    // looks like tooltip, TBD
+
+    if(item._isHidden) {
+      targetItem.extension.push({
+        url: "http://hl7.org/fhir/StructureDefinition/questionnaire-hidden",
+        valueBoolean: true
+      });
+    }
+
+
+    // linkId
+    targetItem.linkId = item.linkId ? item.linkId : item._codePath;
+
+    // Text & prefix
+    targetItem.text = item.question;
+    if (item.prefix) {
+      targetItem.prefix = item.prefix;
+    }
+    // Copy item extensions
+    for (let extField of ['_prefix', '_text']) {
+      let extFieldData = item['obj'+extField];
+      if (extFieldData)
+        targetItem[extField] = extFieldData;
+    }
+
+    // enableWhen
+    if (item.skipLogic) {
+      this._handleSkipLogic(targetItem, item, source)
+    }
+
+    // repeats, handled above
+    // readonly, (editable)
+    if (item.dataType !== "SECTION" && item.dataType !== "TITLE" && item.editable === "0") {
+      targetItem.readOnly = true;
+    }
+
+    this._handleChoiceField(targetItem, item, noExtensions);
+    this._handleTerminologyServer(targetItem, item);
+
+    // initialValue, for default values
+    this._handleInitialValues(targetItem, item);
+    // add LForms Extension to units list. Process units after handling initial values.
+    if (item.units) {
+      this._handleLFormsUnits(targetItem, item);
+    }
+
+    if (item.items && Array.isArray(item.items)) {
+      targetItem.item = [];
+      for (var i=0, iLen=item.items.length; i<iLen; i++) {
+        var newItem = this._processItem(item.items[i], source, noExtensions);
+        targetItem.item.push(newItem);
+      }
+    }
+
+    // the coding instruction is a sub item with a "display" type, and an item-control value as "help"
+    // it is added as a sub item of this item.
+    // http://hl7.org/fhir/StructureDefinition/questionnaire-itemControl, for instructions
+    if (item.codingInstructions) {
+      let helpItem = {
+        "text": item.codingInstructionsPlain ? item.codingInstructionsPlain : item.codingInstructions,
+        "type": "display",
+        "linkId": targetItem.linkId + "-help",
+        "extension": [{
+          "url": "http://hl7.org/fhir/StructureDefinition/questionnaire-itemControl",
+          "valueCodeableConcept": {
+            "text": "Help-Button",
+            "coding": [{
+              "code": "help",
+              "display": "Help-Button",
+              "system": "http://hl7.org/fhir/questionnaire-item-control"
+            }]
+          }
+        }]
+      };
+
+      // format could be 'html' or 'text'
+      if (item.codingInstructionsFormat === 'html') {
+        // add a "_text" field to contain the extension for the string value in the 'text' field
+        // see http://hl7.org/fhir/R4/json.html#primitive
+        helpItem._text = {
+          "extension": [{
+            "url": "http://hl7.org/fhir/StructureDefinition/rendering-xhtml",
+            "valueString": item.codingInstructions
+          }]
+        }
+      }
+
+      if (Array.isArray(targetItem.item)) {
+        targetItem.item.push(helpItem)
+      }
+      else {
+        targetItem.item = [
+          helpItem
+        ]
+      }
+    }
+
+    // handle special constraints for "display" item
+    this._handleSpecialConstraints(targetItem, item);
+
+    // if no extensions are allowed or there is no extension, remove it
+    if (noExtensions || targetItem.extension.length === 0)
+      delete targetItem.extension;
+
+    this.copyFields(item, targetItem, this.itemLevelIgnoredFields);
+    return targetItem
+  },
+
+
+  /**
+   * Process an item's externally defined answer list
+   * @param targetItem an item in FHIR SDC Questionnaire object
+   * @param item an item in the LForms form object
+   * @returns {*}
+   * @private
+   */
+  self._handleExternallyDefined = function(targetItem, item) {
+    if (item.externallyDefined) {
+      targetItem.extension.push({
+        "url": "http://hl7.org/fhir/StructureDefinition/questionnaire-externallydefined",
+        "valueUri": item.externallyDefined
+      });
+    }
+  };
+
+
+  /**
    * Remove repeating items in a form data object
    * @param source a LForms form data object
    * @private
@@ -111,6 +296,15 @@ function addCommonSDCExportFns(ns) {
   self._setFormLevelFields = function(target, source, noExtensions) {
 
     this.copyFields(source, target, this.formLevelFields);
+    // Handle title and name.  In LForms, "name" is the "title", but FHIR
+    // defines both.
+    target.name = source.shortName; // computer friendly
+    target.title = source.name;
+
+    // Handle extensions on title
+    if (source.obj_title)
+      target._title = source.obj_title;
+
     target.code = source.codeList;
 
     // resourceType